+++ internal/app/plugin.go
@@ -13,7 +13,6 @@ import (
 	"github.com/WuKongIM/WuKongIM/internal/runtime/pluginhook"
 	messageusecase "github.com/WuKongIM/WuKongIM/internal/usecase/message"
 	pluginusecase "github.com/WuKongIM/WuKongIM/internal/usecase/plugin"
-	userusecase "github.com/WuKongIM/WuKongIM/internal/usecase/user"
 	pluginhost "github.com/WuKongIM/WuKongIM/pkg/plugin/pluginhost"
 )
 
@@ -86,7 +85,7 @@ func (a *App) wirePluginSubsystem(nodeID uint64) error {
 		Invoker:          invoker,
 		DesiredStore:     pluginDesiredStoreAdapter{store: store},
 		Messages:         pluginMessageSender{app: a},
-		DefaultSenderUID: userusecase.DefaultSystemUID,
+		DefaultSenderUID: a.cfg.Message.SystemUID,
 		SystemUIDs:       a.users,
 		FailOpen:         a.cfg.Plugin.FailOpen,
 		Observer:         a.pluginUsecaseObserver(),